+++ bodhi/server/models.py
@@ -722,11 +722,6 @@ update_bug_table = Table(
     Column('update_id', Integer, ForeignKey('updates.id')),
     Column('bug_id', Integer, ForeignKey('bugs.id')))
 
-user_package_table = Table(
-    'user_package_table', metadata,
-    Column('user_id', Integer, ForeignKey('users.id')),
-    Column('package_id', Integer, ForeignKey('packages.id')))
-
 
 class Release(Base):
     """
@@ -933,13 +928,11 @@ class Package(Base):
         builds (sqlalchemy.orm.collections.InstrumentedList): A list of :class:`Build` objects.
         test_cases (sqlalchemy.orm.collections.InstrumentedList): A list of :class:`TestCase`
             objects.
-        committers (sqlalchemy.orm.collections.InstrumentedList): A list of :class:`User` objects
-            who are committers.
     """
 
     __tablename__ = 'packages'
     __get_by__ = ('name',)
-    __exclude_columns__ = ('id', 'committers', 'test_cases', 'builds',)
+    __exclude_columns__ = ('id', 'test_cases', 'builds',)
 
     name = Column(UnicodeText, nullable=False)
     requirements = Column(UnicodeText)
@@ -947,8 +940,6 @@ class Package(Base):
 
     builds = relationship('Build', backref=backref('package', lazy='joined'))
     test_cases = relationship('TestCase', backref='package', order_by="TestCase.id")
-    committers = relationship('User', secondary=user_package_table,
-                              backref='packages')
 
     __mapper_args__ = {
         'polymorphic_on': type,
@@ -1683,17 +1674,19 @@ class Update(Base):
         We use this as a way to inject an alias into the Update, since it is a required field and
         we don't want callers to have to generate the alias themselves.
         """
-        super(Update, self).__init__(*args, **kwargs)
-
         # Let's give this Update an alias so the DB doesn't become displeased with us.
-        if not self.release:
+        if 'release' not in kwargs:
             raise ValueError('You must specify a Release when creating an Update.')
-        prefix = self.release.id_prefix
+        prefix = kwargs['release'].id_prefix
         year = time.localtime()[0]
         id = hashlib.sha1(str(uuid.uuid4()).encode('utf-8')).hexdigest()[:10]
         alias = u'%s-%s-%s' % (prefix, year, id)
-        log.debug('Setting alias for %s to %s' % (self.get_title(), alias))
         self.alias = alias
+        self.release_id = kwargs['release'].id
+
+        super(Update, self).__init__(*args, **kwargs)
+
+        log.debug('Set alias for %s to %s' % (self.get_title(), alias))
 
     @property
     def side_tag_locked(self):
@@ -3072,12 +3065,7 @@ class Update(Base):
             list: A list of :class:`Users <User>` who have commit access to all of the
                 packages that are contained within this update.
         """
-        people = set([self.user])
-        for build in self.builds:
-            if build.package.committers:
-                for committer in build.package.committers:
-                    people.add(committer)
-        return list(people)
+        return [self.user]
 
     @property
     def product_version(self):
@@ -4068,7 +4056,7 @@ class User(Base):
     """
 
     __tablename__ = 'users'
-    __exclude_columns__ = ('comments', 'updates', 'packages', 'buildroot_overrides')
+    __exclude_columns__ = ('comments', 'updates', 'buildroot_overrides')
     __include_extras__ = ('avatar', 'openid')
     __get_by__ = ('name',)