+++ docs/user/release_notes.rst
@@ -19,6 +19,8 @@ Backwards incompatible changes
   ``True`` instead of ``None`` (:issue:`1921`).
 * Integration with pkgdb is no longer supported (:issue:`1970`).
 * The ``/admin/`` API has been removed (:issue:`1985`).
+* The relationship between Packages and Users was dropped. As a result, the ``packages``
+  parameter in users query API has also been removed (:issue:`1997`).
 * Support for CVE tracking was dropped. It was technically not possible to use the feature, so it
   is unlikely to affect any deployments (:issue:`1998`).
 * The ``processing`` update status has been removed (:issue:`1999`).
@@ -56,7 +58,8 @@ Backwards incompatible changes
   supported in Bodhi 4, even though Bodhi still sends messages similar to the messages it sent in
   the past. Message consumers should not rely on any undocumented fields in these messages. If you
   need information that is not included in the supported schema, please work with the Bodhi project
-  to get the schema adjusted accordingly.
+  to get the schema adjusted accordingly. Bodhi's messages are now documented in
+  :doc:`../server_api/index`.
 
 
 Dependency changes